+++ src/lib.rs
@@ -147,9 +147,6 @@
 //! domains. This seems to matter most on the 1050, which has the widest spread of bank-to-power
 //! domain assignment (according to AN12077).
 //!
-//! There is no support for ECC on 1160 or 1170. The runtime assumes that OCRAM and TCM ECC
-//! is disabled, and that the corresponding memory banks can be used for OCRAM.
-//!
 //! The runtime installs a `cortex-m-rt` `pre_init` function to configure the runtime.
 //! You cannot also define a `pre_init` function, and this crate does not support any
 //! other mechanism for running code before `main()`.